+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ## Added
 
 - User resource.
+- Group resource.
 - Fake 1password storage.
+- Real 1password storage.
 
 [unreleased]: https://github.com/slok/terraform-provider-onepasswordorg/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/slok/terraform-provider-onepasswordorg/releases/tag/v0.1.0